+++ src/EventSubscriber.ts
@@ -189,7 +189,7 @@ class EventSubscriber {
 
     watchRouteFiles.forEach((pattern) =>
       this.subscribeFileWatcher(pattern, () => {
-        loadWithProgress('Loading rails routes', () => this.routes.load());
+        loadWithProgress('Loading rails routes', () => this.routes.load(true));
       })
     );
   }